Create Display_Contacts_Using_Datatable_Apex
Deploy a Component That Lists Contacts
Create a Lightning web component that uses @wire in an Apex controller to retrieve contact records. Display the contact records in a lightning-datatable.

Tip: Remember that JavaScript is a case-sensitive programming language. Be careful to use uppercase and lowercase letters correctly.

Create an Apex controller that retrieves a list of contacts:
Class: ContactController
Method: getContacts()
Fields to query: FirstName, LastName, Email
Create a Lightning web component that displays contacts in a table:
Component name: contactList
Base component: lightning-datatable
Fields to include: FirstName, LastName, Email
Wire the getContacts() method
Add the component to the Working with Data page